more changes

+++ fileUpload.php
@@ -4,13 +4,57 @@ include 'Parser.php';
 
 
 if (isset($_POST['url-submit'])) {
-    
+    try {
+        $temp = tempnam("./temp");
+        file_put_contents($temp, fopen($_POST['url'], 'r'));
+         $parser = new Parser($temp);
+        $parser->parse();
+        $parser->createParseTree();
+        $parser->runValidator();
+        if (file_exists($temp)) {
+                unlink($temp);
+            }
+    } catch (Exception $e) {
+        $_SESSION['errorMessage'] = $e->getMessage();
+        header('Location: uhoh.php');
+    }
+
 }
+
 
 elseif (isset($_POST['file-submit'])) {
     try {
         if ($_FILES["uploaded_file"]["error"] > 0) {
-            echo "Error: " . $_FILES["uploaded_file"]["error"] . "<br />";
+            $errormsg = "";
+            switch ($_FILES["uploaded_file"]["error"]) {
+                case 1:
+                    $errormsg = "File is too big!";
+                    break;
+                case 2:
+                    $errormsg = "File is too  big!";
+                    break;
+                case 3:
+                    $errormsg = "Only part of the file was uplaoded";
+                    break;
+                case 4:
+                    $errormsg = "No file was uploaded";
+                    break;
+                case 6:
+                    $errormsg = "Something went wrong on the server [Error: 6]";
+                    break;
+                case 7:
+                    $errormsg = "Something went wrong on the server [Error: 7]";
+                    break;
+                case 8:
+                    $errormsg = "Something went wrong on the server [Error: 8]";
+                    break;
+                default:
+                    $errormsg = "Something went really wrong!";
+                    break;
+        }
+            
+            $_SESSION['errorMessage'] = $errormsg;
+            header('Location: uhoh.php');
         } else {
             $path = "upload/" . $_FILES["uploaded_file"]["name"];
             move_uploaded_file($_FILES["uploaded_file"]["tmp_name"], $path);
@@ -18,7 +62,9 @@ elseif (isset($_POST['file-submit'])) {
             $parser->parse();
             $parser->createParseTree();
             $parser->runValidator();
-            unlink($path);
+            if (file_exists($path)) {
+                unlink($path);
+            }
         }
     } catch (Exception $e) {
         $_SESSION['errorMessage'] = $e->getMessage;
@@ -27,6 +73,8 @@ elseif (isset($_POST['file-submit'])) {
 } 
 
 elseif (isset($_POST['zip-submit'])) {
+    $info = pathinfo($_POST['uploaded_zip']);
+    
     
 }
 
@@ -39,7 +87,9 @@ elseif (isset($_POST['direct-submit'])) {
         $parser->parse();
         $parser->createParseTree();
         $parser->runValidator();
-        unlink($temp);
+        if (file_exists($temp)) {
+                unlink($temp);
+            }
     } catch (Exception $e) {
         $_SESSION['errorMessage'] = $e->getMessage();
         header('Location: uhoh.php');
